#136 https://github.com/AlibabaRazboinik/habit_tracker/issues/136 тест на удаление привычки не работает

+++ src/main/java/com/example/habitstracker/controllers/HabitApiImpl.java
@@ -1,9 +1,13 @@
 package com.example.habitstracker.controllers;
 
+import com.example.habitstracker.exceptions.HabitPermissionException;
 import com.example.habitstracker.models.Habit;
+import com.example.habitstracker.models.HabitList;
+import com.example.habitstracker.models.UserEntity;
 import com.example.habitstracker.security.UserCredentials;
 import com.example.habitstracker.services.HabitService;
 import com.example.habitstracker.services.MapperService;
+import com.example.habitstracker.services.UserService;
 import com.example.openapi.api.HabitApi;
 import com.example.openapi.dto.HabitDTO;
 import com.example.openapi.dto.IdDTO;
@@ -14,6 +18,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.List;
+
 /**
  * Контроллер для привычки
  */
@@ -22,13 +28,15 @@ public class HabitApiImpl implements HabitApi
 {
     private final HabitService habitService;
     private final MapperService mapperService;
+    private final UserService userService;
     private final Logger log = LoggerFactory.getLogger(this.getClass().getName());
 
     @Autowired
-    public HabitApiImpl(HabitService habitService, MapperService mapperService)
+    public HabitApiImpl(HabitService habitService, MapperService mapperService, UserService userService)
     {
         this.habitService = habitService;
         this.mapperService = mapperService;
+        this.userService = userService;
     }
 
     @Override
@@ -47,9 +55,10 @@ public class HabitApiImpl implements HabitApi
     @Override
     public ResponseEntity<Void> deleteHabit(Long id)
     {
-        UserCredentials credentials = (UserCredentials) SecurityContextHolder.getContext().getAuthentication().getCredentials();
-
         log.info("Delete habit " + id);
+
+        if(isUserHabit(id) == false)
+            throw new HabitPermissionException(id);
 
         habitService.deleteHabit(id);
         return ResponseEntity.ok().build();
@@ -58,7 +67,10 @@ public class HabitApiImpl implements HabitApi
     @Override
     public ResponseEntity<HabitDTO> getHabit(Long id)
     {
-        log.info("Delete habit with id = " + id);
+        log.info("Get habit with id = " + id);
+
+        if(isUserHabit(id) == false)
+            throw new HabitPermissionException(id);
 
         HabitDTO habitDTO = new HabitDTO();
         mapperService.transform(habitService.getHabit(id), habitDTO);
@@ -69,10 +81,23 @@ public class HabitApiImpl implements HabitApi
     public ResponseEntity<Void> updateHabit(Long id, HabitDTO habitDTO)
     {
         log.info("Update habit with id = " + id + " new values " + habitDTO.toInlineString());
+        if(isUserHabit(id) == false)
+            throw new HabitPermissionException(id);
 
         Habit habit = new Habit();
         mapperService.transform(habitDTO, habit);
         habitService.updateHabit(id, habit);
         return ResponseEntity.ok().build();
+    }
+
+    private boolean isUserHabit(Long id){
+        UserCredentials userCredentials = (UserCredentials) SecurityContextHolder.getContext()
+                .getAuthentication().getCredentials();
+        UserEntity user = userService.getById(userCredentials.id());
+        List<Habit> habits = user.getHabitList().getHabits();
+        for(var i = 0; i < habits.size(); i++)
+            if (habits.get(i).getId().equals(id))
+                return true;
+        return false;
     }
 }
